move whitespace to util
