+++ lib/geocoder/phrasematch.js
@@ -4,7 +4,7 @@ const token = require('../text-processing/token');
 const bb = require('../util/bbox');
 const cl = require('../text-processing/closest-lang');
 const filter = require('./filter-sources');
-const whitespace = require('./whitespace');
+const whitespace = require('../util/whitespace');
 const MIN_CORRECTION_LENGTH = require('../constants').MIN_CORRECTION_LENGTH;
 const MAX_QUERY_TOKENS = require('../constants').MAX_QUERY_TOKENS;
 const ENDING_TYPE = require('@mapbox/node-fuzzy-phrase').ENDING_TYPE;
