+++ lib/util/whitespace.js
